Implement consumer authorisation middleware
Implements a middleware (before filter) post authentication to perform “authorisation“ to limit access of consumers to endpoints that are out of their business.

This is a Grape middleware and works by annotating endpoints with an array of `authorised_consumers` (issuer names). If the consumer is not in the list, a forbidden error will be returned.